# backend/docs/openapi.yaml
openapi: "3.1.0"
info:
  title: SecretDrop API
  description: |
    Secure one-time secret sharing API. Secrets are encrypted with
    AES-256-GCM (HKDF-derived keys), stored in SQLite, and permanently
    deleted after a single reveal or expiry.

    ## Security Model

    - The encryption key never touches the server database.
    - Key is carried only in the URL fragment (`#` after the path) which
      browsers and proxies do not send to the server.
    - HKDF binds each key to the recipient's email — the same link cannot
      be opened by a different email address.
    - `recipient_hash` is SHA-256 of the email; the raw email is never stored.

    ## Authentication

    Most endpoints require a JWT Bearer token obtained via OAuth login
    (Google or GitHub). Tokens are short-lived (15 min access, 30 day refresh).
  version: "0.4.0"
  contact:
    name: Bilus Teknoloji
    url: https://github.com/bilustek/secretdrop
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/admin/users:
    get:
      operationId: adminListUsers
      summary: List users (admin)
      description: |
        Returns a paginated list of users with optional search, filter, and sort.
        Requires HTTP Basic Auth (admin credentials).
      security:
        - basicAuth: []
      parameters:
        - name: q
          in: query
          description: Search by email or name (LIKE match)
          schema:
            type: string
        - name: tier
          in: query
          description: Filter by tier
          schema:
            type: string
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            enum: [created_at, email, name, tier, secrets_used]
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersListResponse"
        "401":
          description: Admin authentication required
  /api/v1/admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      summary: Update user (admin)
      description: |
        Updates a user's tier, sets a per-user secrets limit override, or clears
        the override. At least one field must be provided. Requires HTTP Basic Auth.
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserRequest"
            examples:
              change_tier:
                summary: Change user tier
                value:
                  tier: "pro"
              set_override:
                summary: Set per-user secrets limit override
                value:
                  secrets_limit_override: 500
              clear_override:
                summary: Clear per-user secrets limit override
                value:
                  clear_secrets_limit: true
      responses:
        "200":
          description: User updated
        "400":
          description: Invalid request (unknown tier, invalid override value, or no fields)
        "401":
          description: Admin authentication required
        "404":
          description: User not found
  /api/v1/admin/subscriptions:
    get:
      operationId: adminListSubscriptions
      summary: List subscriptions (admin)
      description: |
        Returns a paginated list of subscriptions with user info.
        Requires HTTP Basic Auth.
      security:
        - basicAuth: []
      parameters:
        - name: q
          in: query
          description: Search by user email or name
          schema:
            type: string
        - name: status
          in: query
          description: Filter by subscription status
          schema:
            type: string
            enum: [active, canceled, past_due]
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at, status]
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Paginated subscription list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSubscriptionsListResponse"
        "401":
          description: Admin authentication required
  /api/v1/admin/subscriptions/{id}:
    delete:
      operationId: adminCancelSubscription
      summary: Cancel subscription (admin)
      description: |
        Cancels a subscription in Stripe and updates the database status.
        The user is downgraded to free tier. Requires HTTP Basic Auth.
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Subscription cancelled
        "401":
          description: Admin authentication required
        "404":
          description: Subscription not found
  /api/v1/admin/limits:
    get:
      operationId: adminListLimits
      summary: List tier limits (admin)
      description: |
        Returns all configured tier limits (secrets and recipients per tier).
        Requires HTTP Basic Auth.
      security:
        - basicAuth: []
      responses:
        "200":
          description: Array of tier limits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminLimitsResponse"
              example:
                - tier: "free"
                  secrets_limit: 5
                  recipients_limit: 1
                - tier: "pro"
                  secrets_limit: 100
                  recipients_limit: 10
        "401":
          description: Admin authentication required
  /api/v1/admin/limits/{tier}:
    put:
      operationId: adminUpsertLimits
      summary: Create or update tier limits (admin)
      description: |
        Creates a new tier or updates an existing tier's limits.
        Requires HTTP Basic Auth.
      security:
        - basicAuth: []
      parameters:
        - name: tier
          in: path
          required: true
          description: Tier name (e.g. free, pro, vip)
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpsertLimitsRequest"
            example:
              secrets_limit: 1000
              recipients_limit: 20
      responses:
        "200":
          description: Tier limits created or updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLimitsResponse"
        "400":
          description: Invalid limits (must be positive)
        "401":
          description: Admin authentication required

    delete:
      operationId: adminDeleteLimits
      summary: Delete tier limits (admin)
      description: |
        Deletes a tier from the limits table. The free tier cannot be deleted.
        Users on this tier will fall back to hardcoded defaults.
        Requires HTTP Basic Auth.
      security:
        - basicAuth: []
      parameters:
        - name: tier
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Tier limits deleted
        "400":
          description: Cannot delete the free tier
        "401":
          description: Admin authentication required
        "404":
          description: Tier not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained via OAuth login
    basicAuth:
      type: http
      scheme: basic
      description: Admin credentials (ADMIN_USERNAME / ADMIN_PASSWORD env vars)
  schemas:
...
    AdminUsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserResponse"
        total:
          type: integer
          format: int64
        page:
          type: integer
        per_page:
          type: integer
    AdminUserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        name:
          type: string
        provider:
          type: string
          enum: [google, github]
        tier:
          type: string
        secrets_used:
          type: integer
        secrets_limit:
          type: integer
          description: Effective secrets limit (per-user override > tier limit > fallback)
        secrets_limit_override:
          type: integer
          nullable: true
          description: Per-user secrets limit override (null if using tier default)
        created_at:
          type: string
          format: date-time
    AdminSubscriptionsListResponse:
      type: object
      properties:
        subscriptions:
          type: array
          items:
            $ref: "#/components/schemas/AdminSubscriptionResponse"
        total:
          type: integer
          format: int64
        page:
          type: integer
        per_page:
          type: integer
    AdminSubscriptionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        user_email:
          type: string
          format: email
        user_name:
          type: string
        stripe_customer_id:
          type: string
        stripe_subscription_id:
          type: string
        status:
          type: string
          enum: [active, canceled, past_due]
        current_period_start:
          type: string
          format: date-time
        current_period_end:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    AdminUpdateUserRequest:
      type: object
      description: |
        At least one field must be provided. Fields can be combined
        (e.g., change tier and set override in one request).
      properties:
        tier:
          type: string
          description: New tier for the user (must exist in limits table)
        secrets_limit_override:
          type: integer
          minimum: 1
          description: Per-user secrets limit override (positive integer)
        clear_secrets_limit:
          type: boolean
          description: Set to true to remove the per-user secrets limit override
    AdminLimitsResponse:
      type: object
      properties:
        tier:
          type: string
        secrets_limit:
          type: integer
        recipients_limit:
          type: integer
    AdminUpsertLimitsRequest:
      type: object
      required:
        - secrets_limit
        - recipients_limit
      properties:
        secrets_limit:
          type: integer
          minimum: 1
          description: Maximum number of secrets for this tier
        recipients_limit:
          type: integer
          minimum: 1
          description: Maximum recipients per secret for this tier
...